djangoapp/QUIZ/requirements.txt
django
django-rest-framework
django-rest-knox
django-celery
mysql-client